# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZGRZYT API",
        "description": "Dokumentacja API dla systemu Zgłoszeń i Rejestru Zdarzeń Technicznych (ZGRZYT). Jest to webowa aplikacja do zgłaszania, obsługi i monitorowania błędów w systemach informatycznych.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Logowanie użytkownika",
                "description": "Uwierzytelnia użytkownika na podstawie loginu i hasła, a w odpowiedzi zwraca token dostępowy (Bearer Token) oraz rolę użytkownika. Ten endpoint jest publiczny i nie wymaga autoryzacji.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Dane logowania użytkownika.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "jkowalski"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logowanie pomyślne. Zwraca token i rolę.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|aBcDeFgHiJkLmNoPqRsTuVwXyZ..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd uwierzytelniania (nieprawidłowe dane logowania)."
                    },
                    "422": {
                        "description": "Błąd walidacji (np. brakujące pola)."
                    }
                },
                "security": []
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Tworzenie nowego użytkownika przez Admina/IT",
                "description": "Tworzy nowe konto użytkownika. Dostępne tylko dla uwierzytelnionych użytkowników z rolą 'admin' lub 'it'. Pozwala na zdefiniowanie roli nowego użytkownika.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "login",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jan Kowalski"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "jkowalski"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "j.kowalski@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user",
                                        "enum": [
                                            "user",
                                            "it",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Użytkownik utworzony pomyślnie.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień."
                    },
                    "422": {
                        "description": "Błąd walidacji danych."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/request-account": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Zażądaj utworzenia nowego konta użytkownika",
                "description": "Umożliwia zalogowanemu użytkownikowi wysłanie prośby o utworzenie nowego konta (np. dla nowego pracownika). Tworzy **nieaktywne** konto i automatycznie generuje zgłoszenie do działu IT z prośbą o jego aktywację.",
                "operationId": "requestAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "login",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Anna Nowak"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "anowak"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "a.nowak@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prośba o konto została pomyślnie wysłana.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień."
                    },
                    "422": {
                        "description": "Błąd walidacji danych (np. email już istnieje)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autoryzacja"
                ],
                "summary": "Wylogowanie użytkownika (unieważnienie tokena API)",
                "description": "Unieważnia token dostępowy API (Bearer token), z którym zostało wykonane żądanie. Jest to metoda wylogowania dla klientów API (np. aplikacji frontendowej Vue).",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Token pomyślnie unieważniony.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji (brak lub nieprawidłowy token)."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/logout": {
            "post": {
                "tags": [
                    "Sesja"
                ],
                "summary": "Wylogowanie użytkownika (sesja webowa)",
                "description": "Kończy sesję użytkownika w aplikacji webowej (unieważnia sesję).",
                "operationId": "logoutSession",
                "responses": {
                    "200": {
                        "description": "Pomyślnie wylogowano"
                    },
                    "401": {
                        "description": "Błąd autoryzacji (użytkownik niezalogowany)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model użytkownika",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "login": {
                        "type": "string",
                        "example": "jkowalski"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "j.kowalski@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "it",
                            "admin"
                        ]
                    },
                    "active": {
                        "description": "Status aktywności konta.",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}